+++ CHANGES.md
@@ -5,6 +5,7 @@ Release Notes.
 9.8.0
 ------------------
 
+* Add Spring LDAP 3.3.x-4.x plugin.
 
 All issues and pull requests are [here](https://github.com/apache/skywalking/milestone/263?closed=1)